refactor: rename KottageListItem to KottageListEntry
